move most filter logic to filter components
