+++ src/js/components/Beacon/BeaconQueryUi.js
@@ -19,7 +19,7 @@ const BeaconQueryUi = () => {
   const [filters, setFilters] = useState([]);
   const [form] = Form.useForm();
   const querySections = useAppSelector((state) => state.query.querySections);
-  const maxFilters = useAppSelector((state) => state.config.maxQueryParameters);
+  // const maxFilters = useAppSelector((state) => state.config.maxQueryParameters);
   const beaconUrl = useAppSelector((state) => state.config?.beaconUrl);
 
   const dispatch = useDispatch();
@@ -49,24 +49,12 @@ const BeaconQueryUi = () => {
 
 
   const packageFilters = (values) => {
-    return filters
-      .filter((f) => f.active)
-      .map((f) => ({
+    return filters.filter((f) => f.active).map((f) => ({
         id: values[`filterId${f.index}`],
         operator: "=",
         value: values[`filterValue${f.index}`],
       }));
   };
-
-  const newFilter = (n) => ({ index: `${n}`, active: true});
-  
-
-  const removeFilter = (filter) => {
-    // set to active: false
-    setFilters(filters.map(f => f.index == filter.index? ({ index: filter.index, active: false}) : f))
-  }
-  const activeFilters = filters.filter(f => f.active)
-  const hasMaxFilters = activeFilters.length >= maxFilters;
 
   const packageBeaconJSON = (values) => {
     let query = {};
@@ -110,21 +98,6 @@ const BeaconQueryUi = () => {
     dispatch(makeBeaconQuery(jsonPayload));
   };
 
-  const handleAddFilter = () => {
-    const filterIndex = filters.length + 1;
-    const f = newFilter(filterIndex);
-    setFilters((filters) => [...filters, f]);
-  };
-
-  const handleRemoveFilters = () => {
-    setFilters([]);
-  };
-
-  const handleClearForm = () => {
-    handleRemoveFilters();
-    form.resetFields();
-  };
-
   const wrapperStyle = {
     display: 'flex',
     flexDirection: 'column',
@@ -159,20 +132,7 @@ const BeaconQueryUi = () => {
         <VariantsForm assemblyIdOptions={assemblyIdOptions} form={form} />
         <Form.Item>
           <div style={filterAreaStyle}>
-            <div style={buttonAreaStyle}>
-              <Tooltip title={hasMaxFilters ? `maximum of ${maxFilters} filters permitted` : null}>
-                <Button style={buttonStyle} onClick={handleAddFilter} disabled={hasMaxFilters}>
-                  Add Filter
-                </Button>
-              </Tooltip>
-              <Button style={buttonStyle} onClick={handleRemoveFilters}>
-                Clear Filters
-              </Button>
-              <Button style={{ ...buttonStyle, marginLeft: '20px' }} onClick={handleClearForm}>
-                Clear Form
-              </Button>
-            </div>
-            <Filters filters={filters} form={form} querySections={querySections} removeFilter={removeFilter} />
+            <Filters filters={filters} setFilters={setFilters} form={form} querySections={querySections} />
           </div>
           <Button style={submitButtonStyle} type="primary" htmlType="submit">
             Search Beacon
