+++ src/js/components/Beacon/Filter.js
@@ -2,8 +2,6 @@ import { useEffect, useState } from 'react';
 import { Button, Input, Form, Option, Select, Space, Tooltip} from 'antd';
 import { CloseOutlined, InfoCircleOutlined } from '@ant-design/icons';
 
-
-
 // TODOs:
 // helptext (possibly as tooltip over search key)
 // should be able to remove a particular filter 
@@ -14,8 +12,10 @@ const Filter = ({ filter, form, querySections, removeFilter }) => {
   const handleSelectKey = (_, option) => {
     // set dropdown options for a particular key
     // ie for key "sex", set options to "MALE", "FEMALE", etc
+
     setValueOptions(option.optionsThisKey);
   };
+
 
   // rerender default option when key changes
   useEffect(() => {
@@ -26,7 +26,7 @@ const Filter = ({ filter, form, querySections, removeFilter }) => {
 
   const renderLabel = (searchField) => {
     const units = searchField.config?.units
-    const unitsString = units? `\t(${units})` : ""
+    const unitsString = units? `(${units})` : ""
     return searchField.title + unitsString
   }
 
@@ -46,6 +46,7 @@ const Filter = ({ filter, form, querySections, removeFilter }) => {
  const searchValueOptions = (arr) => {
     return arr.map((v) => ({ label: v, value: v }));
   };
+
 
   return (
     <Space.Compact>
